Extended array kphio testing to subdaily - fixed (experimentally) what subdaily model does with internal daily model and array kphio
